Move license map to BOM object and logically sep. hash functions
