+++ cmd/license.go
@@ -24,7 +24,6 @@ import (
 	"github.com/CycloneDX/sbom-utility/common"
 	"github.com/CycloneDX/sbom-utility/schema"
 	"github.com/CycloneDX/sbom-utility/utils"
-	"github.com/jwangsadinata/go-multimap/slicemultimap"
 	"github.com/spf13/cobra"
 )
 
@@ -92,48 +91,11 @@ type LicenseInfo struct {
 	Service                schema.CDXService       // Do not marshal
 }
 
-// License hashmap
-// Holds resources (e.g., components, services) declared license(s)
-var licenseMap = slicemultimap.New()
+// globals
 var licenseSlice []LicenseInfo
 
 func ClearGlobalLicenseData() {
-	licenseMap.Clear()
 	licenseSlice = nil
-}
-
-func HashLicenseInfo(key string, licenseInfo LicenseInfo, whereFilters []common.WhereFilter) {
-	// Find license usage policy by either license Id, Name or Expression
-	policy, err := FindPolicy(licenseInfo)
-
-	if err != nil {
-		// Show intent to not check for error returns as there is no recovery
-		_ = getLogger().Errorf("%s", MSG_LICENSE_INVALID_DATA)
-		os.Exit(ERROR_VALIDATION)
-	}
-	licenseInfo.License = key
-	licenseInfo.Policy = policy
-	licenseInfo.UsagePolicy = policy.UsagePolicy
-	// Derive values for report filtering
-	licenseInfo.LicenseChoiceType = LC_TYPE_NAMES[licenseInfo.LicenseChoiceTypeValue]
-	licenseInfo.BOMLocation = CDX_LICENSE_LOCATION_NAMES[licenseInfo.BOMLocationValue]
-	licenseSlice = append(licenseSlice, licenseInfo)
-
-	var match bool = true
-	if len(whereFilters) > 0 {
-		mapInfo, _ := utils.ConvertStructToMap(licenseInfo)
-		match, _ = whereFilterMatch(mapInfo, whereFilters)
-	}
-
-	if match {
-		// Hash LicenseInfo by license key (i.e., id|name|expression)
-		licenseMap.Put(key, licenseInfo)
-
-		getLogger().Tracef("Put: %s (`%s`), `%s`)",
-			licenseInfo.ResourceName,
-			licenseInfo.UsagePolicy,
-			licenseInfo.BOMRef)
-	}
 }
 
 func NewCommandLicense() *cobra.Command {
@@ -176,7 +138,7 @@ func licenseCmdImpl(cmd *cobra.Command, args []string) error {
 // 2. (root).metadata.component.licenses[] + all "nested" components
 // 3. (root).components[](.license[]) (each component + all "nested" components)
 // 4. (root).services[](.license[]) (each service + all "nested" services)
-func loadDocumentLicenses(document *schema.BOM, whereFilters []common.WhereFilter) (err error) {
+func loadDocumentLicenses(bom *schema.BOM, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
@@ -184,10 +146,10 @@ func loadDocumentLicenses(document *schema.BOM, whereFilters []common.WhereFilte
 	//licensePolicyConfig.Debug()
 
 	// At this time, fail SPDX format SBOMs as "unsupported" (for "any" format)
-	if !document.FormatInfo.IsCycloneDx() {
+	if !bom.FormatInfo.IsCycloneDx() {
 		err = schema.NewUnsupportedFormatForCommandError(
-			document.GetFilename(),
-			document.FormatInfo.CanonicalName,
+			bom.GetFilename(),
+			bom.FormatInfo.CanonicalName,
 			CMD_LICENSE, FORMAT_ANY)
 		return
 	}
@@ -197,34 +159,34 @@ func loadDocumentLicenses(document *schema.BOM, whereFilters []common.WhereFilte
 
 	// Before looking for license data, fully unmarshal the SBOM
 	// into named structures
-	if err = document.UnmarshalCycloneDXBOM(); err != nil {
+	if err = bom.UnmarshalCycloneDXBOM(); err != nil {
 		return
 	}
 
 	// 1. Hash all licenses in the SBOM metadata (i.e., (root).metadata.component)
 	// Note: this SHOULD represent a summary of all licenses that apply
 	// to the component being described in the SBOM
-	if err = hashMetadataLicenses(document, LC_LOC_METADATA, whereFilters); err != nil {
+	if err = hashMetadataLicenses(bom, LC_LOC_METADATA, whereFilters); err != nil {
 		return
 	}
 
 	// 2. Hash all licenses in (root).metadata.component (+ "nested" components)
-	if err = hashMetadataComponentLicenses(document, LC_LOC_METADATA_COMPONENT, whereFilters); err != nil {
+	if err = hashMetadataComponentLicenses(bom, LC_LOC_METADATA_COMPONENT, whereFilters); err != nil {
 		return
 	}
 
 	// 3. Hash all component licenses found in the (root).components[] (+ "nested" components)
-	pComponents := document.GetCdxComponents()
+	pComponents := bom.GetCdxComponents()
 	if pComponents != nil && len(*pComponents) > 0 {
-		if err = hashComponentsLicenses(*pComponents, LC_LOC_COMPONENTS, whereFilters); err != nil {
+		if err = hashComponentsLicenses(bom, *pComponents, LC_LOC_COMPONENTS, whereFilters); err != nil {
 			return
 		}
 	}
 
 	// 4. Hash all service licenses found in the (root).services[] (array) (+ "nested" services)
-	pServices := document.GetCdxServices()
+	pServices := bom.GetCdxServices()
 	if pServices != nil && len(*pServices) > 0 {
-		if err = hashServicesLicenses(*pServices, LC_LOC_SERVICES, whereFilters); err != nil {
+		if err = hashServicesLicenses(bom, *pServices, LC_LOC_SERVICES, whereFilters); err != nil {
 			return
 		}
 	}
@@ -233,14 +195,14 @@ func loadDocumentLicenses(document *schema.BOM, whereFilters []common.WhereFilte
 }
 
 // Hash the license found in the (root).metadata.licenses[] array
-func hashMetadataLicenses(document *schema.BOM, location int, whereFilters []common.WhereFilter) (err error) {
+func hashMetadataLicenses(bom *schema.BOM, location int, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
-	pLicenses := document.GetCdxMetadataLicenses()
+	pLicenses := bom.GetCdxMetadataLicenses()
 	if pLicenses == nil {
 		sbomError := NewInvalidSBOMError(
-			document,
+			bom,
 			fmt.Sprintf("%s (%s)",
 				MSG_LICENSES_NOT_FOUND,
 				CDX_LICENSE_LOCATION_NAMES[location]),
@@ -262,7 +224,7 @@ func hashMetadataLicenses(document *schema.BOM, location int, whereFilters []com
 		licenseInfo.BOMLocationValue = location
 		licenseInfo.ResourceName = LICENSE_LIST_NOT_APPLICABLE
 		licenseInfo.BOMRef = LICENSE_LIST_NOT_APPLICABLE
-		err = hashLicenseInfoByLicenseType(licenseInfo, whereFilters)
+		err = hashLicenseInfoByLicenseType(bom, licenseInfo, whereFilters)
 
 		if err != nil {
 			return
@@ -273,14 +235,14 @@ func hashMetadataLicenses(document *schema.BOM, location int, whereFilters []com
 }
 
 // Hash the license found in the (root).metadata.component object (and any "nested" components)
-func hashMetadataComponentLicenses(document *schema.BOM, location int, whereFilters []common.WhereFilter) (err error) {
+func hashMetadataComponentLicenses(bom *schema.BOM, location int, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
-	component := document.GetCdxMetadataComponent()
+	component := bom.GetCdxMetadataComponent()
 	if component == nil {
 		sbomError := NewInvalidSBOMError(
-			document,
+			bom,
 			fmt.Sprintf("%s (%s)",
 				MSG_LICENSES_NOT_FOUND,
 				CDX_LICENSE_LOCATION_NAMES[location]),
@@ -293,19 +255,19 @@ func hashMetadataComponentLicenses(document *schema.BOM, location int, whereFilt
 		return
 	}
 
-	_, err = hashComponentLicense(*component, location, whereFilters)
+	_, err = hashComponentLicense(bom, *component, location, whereFilters)
 
 	return
 }
 
 // Hash all licenses found in an array of CDX Components
 // TODO use array of pointer to []CDXComponent
-func hashComponentsLicenses(components []schema.CDXComponent, location int, whereFilters []common.WhereFilter) (err error) {
+func hashComponentsLicenses(bom *schema.BOM, components []schema.CDXComponent, location int, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
 	for _, cdxComponent := range components {
-		_, err = hashComponentLicense(cdxComponent, location, whereFilters)
+		_, err = hashComponentLicense(bom, cdxComponent, location, whereFilters)
 		if err != nil {
 			return
 		}
@@ -315,12 +277,12 @@ func hashComponentsLicenses(components []schema.CDXComponent, location int, wher
 
 // Hash all licenses found in an array of CDX Services
 // TODO use array of pointer to []CDXService
-func hashServicesLicenses(services []schema.CDXService, location int, whereFilters []common.WhereFilter) (err error) {
+func hashServicesLicenses(bom *schema.BOM, services []schema.CDXService, location int, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
 	for _, cdxServices := range services {
-		err = hashServiceLicense(cdxServices, location, whereFilters)
+		err = hashServiceLicense(bom, cdxServices, location, whereFilters)
 		if err != nil {
 			return
 		}
@@ -329,7 +291,7 @@ func hashServicesLicenses(services []schema.CDXService, location int, whereFilte
 }
 
 // Hash a CDX Component's licenses and recursively those of any "nested" components
-func hashComponentLicense(cdxComponent schema.CDXComponent, location int, whereFilters []common.WhereFilter) (li *LicenseInfo, err error) {
+func hashComponentLicense(bom *schema.BOM, cdxComponent schema.CDXComponent, location int, whereFilters []common.WhereFilter) (li *LicenseInfo, err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 	var licenseInfo LicenseInfo
@@ -347,7 +309,7 @@ func hashComponentLicense(cdxComponent schema.CDXComponent, location int, whereF
 			if cdxComponent.BOMRef != nil {
 				licenseInfo.BOMRef = *cdxComponent.BOMRef
 			}
-			err = hashLicenseInfoByLicenseType(licenseInfo, whereFilters)
+			err = hashLicenseInfoByLicenseType(bom, licenseInfo, whereFilters)
 
 			if err != nil {
 				// Show intent to not check for error returns as there no intent to recover
@@ -364,7 +326,7 @@ func hashComponentLicense(cdxComponent schema.CDXComponent, location int, whereF
 		if cdxComponent.BOMRef != nil {
 			licenseInfo.BOMRef = *cdxComponent.BOMRef
 		}
-		HashLicenseInfo(LICENSE_NO_ASSERTION, licenseInfo, whereFilters)
+		HashLicenseInfo(bom, LICENSE_NO_ASSERTION, licenseInfo, whereFilters)
 
 		getLogger().Warningf("%s: %s (name:`%s`, version: `%s`, package-url: `%s`)",
 			"No license found for component. bomRef",
@@ -379,7 +341,7 @@ func hashComponentLicense(cdxComponent schema.CDXComponent, location int, whereF
 	// Recursively hash licenses for all child components (i.e., hierarchical composition)
 	pComponents := cdxComponent.Components
 	if pComponents != nil && len(*pComponents) > 0 {
-		err = hashComponentsLicenses(*pComponents, location, whereFilters)
+		err = hashComponentsLicenses(bom, *pComponents, location, whereFilters)
 		if err != nil {
 			return
 		}
@@ -390,7 +352,7 @@ func hashComponentLicense(cdxComponent schema.CDXComponent, location int, whereF
 
 // Hash all licenses found in a CDX Service
 // TODO use pointer to CDXService
-func hashServiceLicense(cdxService schema.CDXService, location int, whereFilters []common.WhereFilter) (err error) {
+func hashServiceLicense(bom *schema.BOM, cdxService schema.CDXService, location int, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
@@ -408,7 +370,7 @@ func hashServiceLicense(cdxService schema.CDXService, location int, whereFilters
 				licenseInfo.BOMRef = *cdxService.BOMRef
 			}
 			licenseInfo.BOMLocationValue = location
-			err = hashLicenseInfoByLicenseType(licenseInfo, whereFilters)
+			err = hashLicenseInfoByLicenseType(bom, licenseInfo, whereFilters)
 
 			if err != nil {
 				return
@@ -423,7 +385,7 @@ func hashServiceLicense(cdxService schema.CDXService, location int, whereFilters
 		if cdxService.BOMRef != nil {
 			licenseInfo.BOMRef = *cdxService.BOMRef
 		}
-		HashLicenseInfo(LICENSE_NO_ASSERTION, licenseInfo, whereFilters)
+		HashLicenseInfo(bom, LICENSE_NO_ASSERTION, licenseInfo, whereFilters)
 
 		getLogger().Warningf("%s: %s (name: `%s`, version: `%s`)",
 			"No license found for service. bomRef",
@@ -438,7 +400,7 @@ func hashServiceLicense(cdxService schema.CDXService, location int, whereFilters
 	// Recursively hash licenses for all child components (i.e., hierarchical composition)
 	pServices := cdxService.Services
 	if pServices != nil && len(*pServices) > 0 {
-		err = hashServicesLicenses(*pServices, location, whereFilters)
+		err = hashServicesLicenses(bom, *pServices, location, whereFilters)
 		if err != nil {
 			// Show intent to not check for error returns as there is no recovery
 			_ = getLogger().Errorf("Unable to hash empty license: %v", licenseInfo)
@@ -454,7 +416,7 @@ func hashServiceLicense(cdxService schema.CDXService, location int, whereFilters
 // 2. Where the license was found within the SBOM
 // 3. The entity name (e.g., service or component name) that declared the license
 // 4. The entity local BOM reference (i.e., "bomRef")
-func hashLicenseInfoByLicenseType(licenseInfo LicenseInfo, whereFilters []common.WhereFilter) (err error) {
+func hashLicenseInfoByLicenseType(bom *schema.BOM, licenseInfo LicenseInfo, whereFilters []common.WhereFilter) (err error) {
 	getLogger().Enter()
 	defer getLogger().Exit(err)
 
@@ -463,13 +425,13 @@ func hashLicenseInfoByLicenseType(licenseInfo LicenseInfo, whereFilters []common
 
 	if licenseChoice.License.Id != "" {
 		licenseInfo.LicenseChoiceTypeValue = LC_TYPE_ID
-		HashLicenseInfo(licenseChoice.License.Id, licenseInfo, whereFilters)
+		HashLicenseInfo(bom, licenseChoice.License.Id, licenseInfo, whereFilters)
 	} else if licenseChoice.License.Name != "" {
 		licenseInfo.LicenseChoiceTypeValue = LC_TYPE_NAME
-		HashLicenseInfo(licenseChoice.License.Name, licenseInfo, whereFilters)
+		HashLicenseInfo(bom, licenseChoice.License.Name, licenseInfo, whereFilters)
 	} else if licenseChoice.Expression != "" {
 		licenseInfo.LicenseChoiceTypeValue = LC_TYPE_EXPRESSION
-		HashLicenseInfo(licenseChoice.Expression, licenseInfo, whereFilters)
+		HashLicenseInfo(bom, licenseChoice.Expression, licenseInfo, whereFilters)
 	} else {
 		// Note: This code path only executes if hashing is performed
 		// without schema validation (which would find this as an error)
@@ -482,4 +444,38 @@ func hashLicenseInfoByLicenseType(licenseInfo LicenseInfo, whereFilters []common
 	}
 
 	return
+}
+
+func HashLicenseInfo(bom *schema.BOM, key string, licenseInfo LicenseInfo, whereFilters []common.WhereFilter) {
+	// Find license usage policy by either license Id, Name or Expression
+	policy, err := FindPolicy(licenseInfo)
+
+	if err != nil {
+		// Show intent to not check for error returns as there is no recovery
+		_ = getLogger().Errorf("%s", MSG_LICENSE_INVALID_DATA)
+		os.Exit(ERROR_VALIDATION)
+	}
+	licenseInfo.License = key
+	licenseInfo.Policy = policy
+	licenseInfo.UsagePolicy = policy.UsagePolicy
+	// Derive values for report filtering
+	licenseInfo.LicenseChoiceType = LC_TYPE_NAMES[licenseInfo.LicenseChoiceTypeValue]
+	licenseInfo.BOMLocation = CDX_LICENSE_LOCATION_NAMES[licenseInfo.BOMLocationValue]
+	licenseSlice = append(licenseSlice, licenseInfo)
+
+	var match bool = true
+	if len(whereFilters) > 0 {
+		mapInfo, _ := utils.ConvertStructToMap(licenseInfo)
+		match, _ = whereFilterMatch(mapInfo, whereFilters)
+	}
+
+	if match {
+		// Hash LicenseInfo by license key (i.e., id|name|expression)
+		bom.LicenseMap.Put(key, licenseInfo)
+
+		getLogger().Tracef("Put: %s (`%s`), `%s`)",
+			licenseInfo.ResourceName,
+			licenseInfo.UsagePolicy,
+			licenseInfo.BOMRef)
+	}
 }
